refactor dreamzero stages
