--- python/sglang/multimodal_gen/runtime/pipelines_core/stages/dreamzero/__init__.py
@@ -1,24 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-from sglang.multimodal_gen.runtime.pipelines_core.stages.dreamzero.action_postproc import (
-    DreamZeroActionOutputStage,
-)
-from sglang.multimodal_gen.runtime.pipelines_core.stages.dreamzero.denoising import (
-    DreamZeroCausalDenoisingStage,
-)
-from sglang.multimodal_gen.runtime.pipelines_core.stages.dreamzero.image_encoding import (
-    DreamZeroVisualEncodingStage,
-)
-from sglang.multimodal_gen.runtime.pipelines_core.stages.dreamzero.obs_prep import (
-    DreamZeroObsPrepStage,
-)
-from sglang.multimodal_gen.runtime.pipelines_core.stages.dreamzero.text_encoding import (
-    DreamZeroTextEncodingStage,
-)
-
-__all__ = [
-    "DreamZeroActionOutputStage",
-    "DreamZeroCausalDenoisingStage",
-    "DreamZeroObsPrepStage",
-    "DreamZeroTextEncodingStage",
-    "DreamZeroVisualEncodingStage",
-]
