+++ python/sglang/multimodal_gen/runtime/pipelines_core/stages/model_specific_stages/dreamzero/image_encoding.py
@@ -1,16 +1,20 @@
 # SPDX-License-Identifier: Apache-2.0
 from __future__ import annotations
 
+import time
+from collections.abc import Mapping
 from typing import Any
 
+import numpy as np
 import torch
 import torch.nn.functional as F
 
-from sglang.multimodal_gen.runtime.managers.dreamzero_session_cache import (
+from sglang.multimodal_gen.runtime.pipelines_core.stages.model_specific_stages.dreamzero.session_cache import (
     DreamZeroCachePool,
     DreamZeroCachePoolManager,
     DreamZeroRequestCache,
-    enter_request_cache,
+    normalize_batched_session_fields,
+    record_session_timing,
 )
 from sglang.multimodal_gen.runtime.managers.forward_context import set_forward_context
 from sglang.multimodal_gen.runtime.managers.memory_managers.component_manager import (
@@ -21,8 +25,8 @@ from sglang.multimodal_gen.runtime.pipelines_core.stages.base import PipelineSta
 from sglang.multimodal_gen.runtime.pipelines_core.stages.validators import (
     VerificationResult,
 )
-from sglang.multimodal_gen.runtime.pipelines_core.stages.dreamzero.utils import (
-    infer_dreamzero_batch_size,
+from sglang.multimodal_gen.runtime.pipelines_core.stages.model_specific_stages.dreamzero.utils import (
+    infer_dreamzero_model_input_batch_size,
 )
 from sglang.multimodal_gen.runtime.platforms import current_platform
 from sglang.multimodal_gen.runtime.server_args import ServerArgs
@@ -108,6 +112,66 @@ def _dreamzero_videos(batch: Req) -> torch.Tensor:
     if videos_input is None:
         raise ValueError("DreamZero image encoding requires 'images' or 'videos'")
     return _as_bcthw(videos_input)
+
+
+class DreamZeroObsPrepStage(PipelineStage):
+    """Prepare DreamZero request-local observation inputs."""
+
+    def verify_output(self, batch: Req, server_args: ServerArgs) -> VerificationResult:
+        result = VerificationResult()
+        result.add_check(
+            "dreamzero_inputs",
+            batch.dreamzero_inputs,
+            lambda value: isinstance(value, dict),
+        )
+        return result
+
+    @staticmethod
+    def _to_tensor_tree(value: Any) -> Any:
+        if isinstance(value, np.ndarray):
+            return torch.from_numpy(value)
+        if isinstance(value, Mapping):
+            return {
+                key: DreamZeroObsPrepStage._to_tensor_tree(item)
+                for key, item in value.items()
+            }
+        if isinstance(value, list):
+            return [DreamZeroObsPrepStage._to_tensor_tree(item) for item in value]
+        if isinstance(value, tuple):
+            return tuple(DreamZeroObsPrepStage._to_tensor_tree(item) for item in value)
+        return value
+
+    def forward(self, batch: Req, server_args: ServerArgs) -> Req:
+        normalized_input = batch.extra.get("dreamzero_normalized_input")
+        if normalized_input is None:
+            raise ValueError(
+                "DreamZero request requires 'dreamzero_normalized_input' in Req.extra"
+            )
+        if not isinstance(normalized_input, Mapping):
+            raise TypeError("DreamZero normalized input must be a mapping")
+
+        model_inputs = self._to_tensor_tree(dict(normalized_input))
+        normalize_start = time.perf_counter()
+        if not server_args.pipeline_config.disable_autocast:
+            for key, value in list(model_inputs.items()):
+                if torch.is_tensor(value) and value.dtype == torch.float32:
+                    model_inputs[key] = value.to(dtype=torch.bfloat16)
+
+        batch.dreamzero_inputs = model_inputs
+        batch_size = infer_dreamzero_model_input_batch_size(model_inputs)
+        session_ids, reset_mask = normalize_batched_session_fields(
+            session_ids=batch.extra.get("dreamzero_session_ids"),
+            reset_mask=batch.extra.get("dreamzero_reset_mask"),
+            batch_size=batch_size,
+        )
+        batch.dreamzero_session_ids = session_ids
+        batch.dreamzero_reset_mask = reset_mask
+        record_session_timing(
+            batch,
+            "session_normalize_ms",
+            (time.perf_counter() - normalize_start) * 1000,
+        )
+        return batch
 
 
 class DreamZeroVisualEncodingStage(PipelineStage):
@@ -430,23 +494,7 @@ class DreamZeroVisualEncodingStage(PipelineStage):
                 return self._normalize_sglang_wan_latent(vae, posterior).to(dtype=dtype)
 
     def forward(self, batch: Req, server_args: ServerArgs) -> Req:
-        arch = server_args.pipeline_config.dit_config.arch_config
-        max_chunk_size = int(getattr(arch, "max_chunk_size", -1))
-        local_attn_size = (
-            -1
-            if max_chunk_size == -1
-            else max_chunk_size * int(arch.num_frame_per_block) + 1
-        )
-        inputs: dict[str, Any] = batch.dreamzero_inputs
-        request_cache, _ = enter_request_cache(
-            batch,
-            self.cache_manager,
-            local_attn_size=local_attn_size,
-            batch_size=infer_dreamzero_batch_size(
-                inputs,
-                error_message="DreamZero visual stage cannot infer batch size",
-            ),
-        )
+        request_cache: DreamZeroRequestCache = batch.dreamzero_cache
         return self._forward_cache_manager(batch, server_args, request_cache)
 
     def _forward_cache_manager(
